+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * 【core   】     Img.scale缩小默认使用平滑模式，增加scale方法重载可选模式（issue#I4MY6X@Gitee）
 * 【core   】     excel添加写入图片的方法（pr#486@Gitee）
 * 【core   】     增加CollStreamUtil.groupBy（pr#484@Gitee）
+* 【core   】     增加CollUtil.setValueByMap（pr#482@Gitee）
 * 
 ### 🐞Bug修复
 * 【core   】     LineReadWatcher#onModify文件清空判断问题（issue#2013@Github）